--- client/src/App.js
@@ -1,31 +1,0 @@
-import React, { useState, useEffect } from "react";
-import logo from "./logo.svg";
-import "./App.css";
-import axios from "axios";
-
-const App = () => {
-  const [message, setMessage] = useState("");
-
-  const getMessage = async () => {
-    const res = await axios.get(`/api/`);
-    console.log(res.data);
-    setMessage(res.data);
-  };
-
-  useEffect(() => {
-    (async () => {
-      await getMessage();
-    })();
-  }, []);
-
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>{message}</p>
-      </header>
-    </div>
-  );
-};
-
-export default App;
